requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
scipy>=1.10.0
filterpy>=1.4.5
PyYAML>=6.0
pygame>=2.5.0
loguru>=0.7.0
lap>=0.4.0
